implmentation different page + component
